+++ packages/echarts/CHANGELOG.md
@@ -1,3 +1,11 @@
+/*
+ * SPDX-FileCopyrightText: 2026 Siemens AG
+ * 
+ * SPDX-License-Identifier: MIT
+ * 
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
 # @siemens/ix-echarts
 
 ## 3.0.1